Update binary_pgm_1.py
"create()": This function creates a new binary file named "Items.dat" and allows the user to add multiple records to the file, each consisting of an item number, name, price, and category. The function uses the "pickle" library to write the records to the file in binary format.

"add()": This function allows the user to add more records to the existing "Items.dat" binary file. It uses the same process as the "create()" function.

"modify()": This function allows the user to modify the price of an item in the "Items.dat" binary file. The user is prompted to enter the item number of the item they wish to modify and the new price. The function uses the "pickle" library to read the records in the file and update the price of the specified item.

"display_old()": This function allows the user to view the records in the original "Items.dat" file. It uses the "pickle" library to read the records and print them to the console.

"delete_particular()": This function allows the user to delete a specific record from the "Items.dat" file. The user is prompted to enter the item number of the record they wish to delete. A new binary file named "Itemsnew.dat" is created, and all the records except the specified one are written to this new file.

"display_new()": This function allows the user to view the records in the new binary file named "Itemsnew.dat". It uses the "pickle" library to read the records and print them to the console.

The program also has a menu that allows the user to select which function to perform. The program runs in a while loop until the user chooses to exit.